+++ R/cusum.R
@@ -107,7 +107,10 @@ cusum.mds_ts <- function(
   # Set analysis_of
   if (is.na(analysis_of)){
     name <- paste(names(ts_event), "of",
-                  paste(attributes(df)$nLabels$nA, collapse=" for "))
+                  paste0(attributes(df)$analysis$device_level_source, " ",
+                         attributes(df)$analysis$device_level, ":",
+                         attributes(df)$analysis$event_level_source, " ",
+                         attributes(df)$analysis$event_level))
   } else name <- analysis_of
 
   out <- data.frame(time=df$time,
@@ -128,10 +131,13 @@ cusum.default <- function(
 ){
   input_param_checker(df, "data.frame")
   input_param_checker(c("time", "event"), check_names=df)
-  input_param_checker(eval_period, "integer")
+  input_param_checker(eval_period, "numeric", null_ok=T, max_length=1)
   input_param_checker(delta, "numeric", null_ok=F, max_length=1)
   input_param_checker(H, "numeric", null_ok=F, max_length=1)
   input_param_checker(zero_rate, "numeric", null_ok=F, max_length=1)
+  if (!is.null(eval_period)){
+    if (eval_period %% 1 != 0) stop("eval_period must be an integer")
+  }
   if (delta <= 0) stop("delta must be >0")
   if (!is.null(H)){ if(H <= 0) stop("H must be >0")}
   if (zero_rate < 0 | zero_rate > 1) stop("zero_rate must be in range [0, 1]")