+++ R/shewhart.R
@@ -110,7 +110,10 @@ shewhart.mds_ts <- function(
   # Set analysis_of
   if (is.na(analysis_of)){
     name <- paste(names(ts_event), "of",
-                  paste(attributes(df)$nLabels$nA, collapse=" for "))
+                  paste0(attributes(df)$analysis$device_level_source, " ",
+                         attributes(df)$analysis$device_level, ":",
+                         attributes(df)$analysis$event_level_source, " ",
+                         attributes(df)$analysis$event_level))
   } else name <- analysis_of
 
   out <- data.frame(time=df$time,
@@ -130,9 +133,12 @@ shewhart.default <- function(
 ){
   input_param_checker(df, "data.frame")
   input_param_checker(c("time", "event"), check_names=df)
-  input_param_checker(eval_period, "integer")
   input_param_checker(zero_rate, "numeric", null_ok=F, max_length=1)
   input_param_checker(we_rule, "integer", null_ok=F, max_length=1)
+  input_param_checker(eval_period, "numeric", null_ok=T, max_length=1)
+  if (!is.null(eval_period)){
+    if (eval_period %% 1 != 0) stop("eval_period must be an integer")
+  }
   if (zero_rate < 0 | zero_rate > 1) stop("zero_rate must be in range [0, 1]")
   if (we_rule < 1 | we_rule > 4) stop("we_rule must be in range [1L, 4L]")